requirements.txt
datasets==2.19.2
huggingface_hub==0.23.3
opencv-contrib-python==4.10.0.82
Pillow==10.3.0
paddlepaddle==2.6.1

scikit-image==0.23.2
imgaug==0.4.0
pyclipper==1.3.0.post5
lmdb==1.4.1
torch==2.3.1
torchvision==0.18.1
einops==0.8.0
imutils==0.5.4
Unidecode==1.3.8
pandas==2.2.2